Mixin for Semisupervised training
