+++ src/scvi/external/methylvi/_model.py
@@ -6,7 +6,6 @@ from collections import defaultdict
 from functools import partial
 from typing import TYPE_CHECKING
 
-from scvi.data._utils import get_anndata_attribute
 from scvi.external.methylvi import METHYLVI_REGISTRY_KEYS
 
 if TYPE_CHECKING:
@@ -26,24 +25,19 @@ import torch
 from scvi import REGISTRY_KEYS, settings
 from scvi.data import AnnDataManager, fields
 from scvi.data._constants import _SETUP_ARGS_KEY
-from scvi.dataloaders import SemiSupervisedDataSplitter
 from scvi.external.methylvi._utils import _context_cov_key, _context_mc_key
-from scvi.model._utils import (
-    get_max_epochs_heuristic,
-)
 from scvi.model.base import (
     ArchesMixin,
     BaseModelClass,
+    SemiSupervisedMixin,
     UnsupervisedTrainingMixin,
     VAEMixin,
 )
 from scvi.model.base._de_core import (
     _de_core,
 )
-from scvi.train import SemiSupervisedTrainingPlan, TrainRunner
-from scvi.train._callbacks import SubSampleLabels
+from scvi.train import SemiSupervisedTrainingPlan
 from scvi.utils import setup_anndata_dsp
-from scvi.utils._docstrings import devices_dsp
 
 from ._module import METHYLANVAE, METHYLVAE
 from ._utils import scmc_raw_counts_properties
@@ -633,7 +627,7 @@ class METHYLVI(VAEMixin, UnsupervisedTrainingMixin, ArchesMixin, BaseModelClass)
         return result
 
 
-class METHYLANVI(VAEMixin, ArchesMixin, BaseModelClass):
+class METHYLANVI(VAEMixin, SemiSupervisedMixin, ArchesMixin, BaseModelClass):
     """Methylation annotation using variational inference :cite:p:`Weinberger23`.
 
     Inspired from M1 + M2 model, as described in (https://arxiv.org/pdf/1406.5298.pdf).
@@ -744,25 +738,6 @@ class METHYLANVI(VAEMixin, ArchesMixin, BaseModelClass):
         self.was_pretrained = False
         self.n_labels = n_labels
 
-    def _set_indices_and_labels(self):
-        """Set indices for labeled and unlabeled cells."""
-        labels_state_registry = self.adata_manager.get_state_registry(REGISTRY_KEYS.LABELS_KEY)
-        self.original_label_key = labels_state_registry.original_key
-        self.unlabeled_category_ = labels_state_registry.unlabeled_category
-
-        labels = get_anndata_attribute(
-            self.adata,
-            self.adata_manager.data_registry.labels.attr_name,
-            self.original_label_key,
-            mod_key=self.adata_manager.data_registry.labels.mod_key,
-        ).ravel()
-        self._label_mapping = labels_state_registry.categorical_mapping
-
-        # set unlabeled and labeled indices
-        self._unlabeled_indices = np.argwhere(labels == self.unlabeled_category_).ravel()
-        self._labeled_indices = np.argwhere(labels != self.unlabeled_category_).ravel()
-        self._code_to_label = dict(enumerate(self._label_mapping))
-
     def predict(
         self,
         mdata: MuData | None = None,
@@ -833,100 +808,6 @@ class METHYLANVI(VAEMixin, ArchesMixin, BaseModelClass):
                 index=mdata.obs_names[indices],
             )
             return pred
-
-    @devices_dsp.dedent
-    def train(
-        self,
-        max_epochs: int | None = None,
-        n_samples_per_label: float | None = None,
-        check_val_every_n_epoch: int | None = None,
-        train_size: float = 0.9,
-        validation_size: float | None = None,
-        shuffle_set_split: bool = True,
-        batch_size: int = 128,
-        accelerator: str = "auto",
-        devices: int | list[int] | str = "auto",
-        datasplitter_kwargs: dict | None = None,
-        plan_kwargs: dict | None = None,
-        **trainer_kwargs,
-    ):
-        """Train the model.
-
-        Parameters
-        ----------
-        max_epochs
-            Number of passes through the dataset for semisupervised training.
-        n_samples_per_label
-            Number of subsamples for each label class to sample per epoch. By default, there
-            is no label subsampling.
-        check_val_every_n_epoch
-            Frequency with which metrics are computed on the data for validation set for both
-            the unsupervised and semisupervised trainers. If you'd like a different frequency for
-            the semisupervised trainer, set check_val_every_n_epoch in semisupervised_train_kwargs.
-        train_size
-            Size of training set in the range [0.0, 1.0].
-        validation_size
-            Size of the test set. If `None`, defaults to 1 - `train_size`. If
-            `train_size + validation_size < 1`, the remaining cells belong to a test set.
-        shuffle_set_split
-            Whether to shuffle indices before splitting. If `False`, the val, train,
-            and test set are split in the sequential order of the data according to
-            `validation_size` and `train_size` percentages.
-        batch_size
-            Minibatch size to use during training.
-        %(param_accelerator)s
-        %(param_devices)s
-        datasplitter_kwargs
-            Additional keyword arguments passed into
-            :class:`~scvi.dataloaders.SemiSupervisedDataSplitter`.
-        plan_kwargs
-            Keyword args for :class:`~scvi.train.SemiSupervisedTrainingPlan`. Keyword
-            arguments passed to `train()` will overwrite values present in `plan_kwargs`,
-            when appropriate.
-        **trainer_kwargs
-            Other keyword args for :class:`~scvi.train.Trainer`.
-        """
-        if max_epochs is None:
-            max_epochs = get_max_epochs_heuristic(self.adata.n_obs)
-
-            if self.was_pretrained:
-                max_epochs = int(np.min([10, np.max([2, round(max_epochs / 3.0)])]))
-
-        logger.info(f"Training for {max_epochs} epochs.")
-
-        plan_kwargs = {} if plan_kwargs is None else plan_kwargs
-        datasplitter_kwargs = datasplitter_kwargs or {}
-
-        # if we have labeled cells, we want to subsample labels each epoch
-        sampler_callback = [SubSampleLabels()] if len(self._labeled_indices) != 0 else []
-
-        data_splitter = SemiSupervisedDataSplitter(
-            adata_manager=self.adata_manager,
-            train_size=train_size,
-            validation_size=validation_size,
-            shuffle_set_split=shuffle_set_split,
-            n_samples_per_label=n_samples_per_label,
-            batch_size=batch_size,
-            **datasplitter_kwargs,
-        )
-        training_plan = self._training_plan_cls(self.module, self.n_labels, **plan_kwargs)
-
-        if "callbacks" in trainer_kwargs.keys():
-            trainer_kwargs["callbacks"] + [sampler_callback]
-        else:
-            trainer_kwargs["callbacks"] = sampler_callback
-
-        runner = TrainRunner(
-            self,
-            training_plan=training_plan,
-            data_splitter=data_splitter,
-            max_epochs=max_epochs,
-            accelerator=accelerator,
-            devices=devices,
-            check_val_every_n_epoch=check_val_every_n_epoch,
-            **trainer_kwargs,
-        )
-        return runner()
 
     @classmethod
     @setup_anndata_dsp.dedent
