+++ src/scvi/model/_scanvi.py
@@ -15,7 +15,7 @@ from scvi.data._constants import (
     _SETUP_ARGS_KEY,
     ADATA_MINIFY_TYPE,
 )
-from scvi.data._utils import _get_adata_minify_type, _is_minified, get_anndata_attribute
+from scvi.data._utils import _get_adata_minify_type, _is_minified
 from scvi.data.fields import (
     CategoricalJointObsField,
     CategoricalObsField,
@@ -24,15 +24,18 @@ from scvi.data.fields import (
     NumericalJointObsField,
     NumericalObsField,
 )
-from scvi.dataloaders import SemiSupervisedDataSplitter
-from scvi.model._utils import _init_library_size, get_max_epochs_heuristic
+from scvi.model._utils import _init_library_size
 from scvi.module import SCANVAE
-from scvi.train import SemiSupervisedTrainingPlan, TrainRunner
-from scvi.train._callbacks import SubSampleLabels
+from scvi.train import SemiSupervisedTrainingPlan
 from scvi.utils import setup_anndata_dsp
-from scvi.utils._docstrings import devices_dsp
 
-from .base import ArchesMixin, BaseMinifiedModeModelClass, RNASeqMixin, VAEMixin
+from .base import (
+    ArchesMixin,
+    BaseMinifiedModeModelClass,
+    RNASeqMixin,
+    SemisupervisedTrainingMixin,
+    VAEMixin,
+)
 
 if TYPE_CHECKING:
     from collections.abc import Sequence
@@ -45,7 +48,9 @@ if TYPE_CHECKING:
 logger = logging.getLogger(__name__)
 
 
-class SCANVI(RNASeqMixin, VAEMixin, ArchesMixin, BaseMinifiedModeModelClass):
+class SCANVI(
+    RNASeqMixin, SemisupervisedTrainingMixin, VAEMixin, ArchesMixin, BaseMinifiedModeModelClass
+):
     """Single-cell annotation using variational inference :cite:p:`Xu21`.
 
     Inspired from M1 + M2 model, as described in (https://arxiv.org/pdf/1406.5298.pdf).
@@ -246,24 +251,6 @@ class SCANVI(RNASeqMixin, VAEMixin, ArchesMixin, BaseMinifiedModeModelClass):
 
         return scanvi_model
 
-    def _set_indices_and_labels(self):
-        """Set indices for labeled and unlabeled cells."""
-        labels_state_registry = self.adata_manager.get_state_registry(REGISTRY_KEYS.LABELS_KEY)
-        self.original_label_key = labels_state_registry.original_key
-        self.unlabeled_category_ = labels_state_registry.unlabeled_category
-
-        labels = get_anndata_attribute(
-            self.adata,
-            self.adata_manager.data_registry.labels.attr_name,
-            self.original_label_key,
-        ).ravel()
-        self._label_mapping = labels_state_registry.categorical_mapping
-
-        # set unlabeled and labeled indices
-        self._unlabeled_indices = np.argwhere(labels == self.unlabeled_category_).ravel()
-        self._labeled_indices = np.argwhere(labels != self.unlabeled_category_).ravel()
-        self._code_to_label = dict(enumerate(self._label_mapping))
-
     def predict(
         self,
         adata: AnnData | None = None,
@@ -338,98 +325,6 @@ class SCANVI(RNASeqMixin, VAEMixin, ArchesMixin, BaseMinifiedModeModelClass):
                 index=adata.obs_names[indices],
             )
             return pred
-
-    @devices_dsp.dedent
-    def train(
-        self,
-        max_epochs: int | None = None,
-        n_samples_per_label: float | None = None,
-        check_val_every_n_epoch: int | None = None,
-        train_size: float | None = None,
-        validation_size: float | None = None,
-        shuffle_set_split: bool = True,
-        batch_size: int = 128,
-        accelerator: str = "auto",
-        devices: int | list[int] | str = "auto",
-        datasplitter_kwargs: dict | None = None,
-        plan_kwargs: dict | None = None,
-        **trainer_kwargs,
-    ):
-        """Train the model.
-
-        Parameters
-        ----------
-        max_epochs
-            Number of passes through the dataset for semisupervised training.
-        n_samples_per_label
-            Number of subsamples for each label class to sample per epoch. By default, there
-            is no label subsampling.
-        check_val_every_n_epoch
-            Frequency with which metrics are computed on the data for validation set for both
-            the unsupervised and semisupervised trainers. If you'd like a different frequency for
-            the semisupervised trainer, set check_val_every_n_epoch in semisupervised_train_kwargs.
-        train_size
-            Size of training set in the range [0.0, 1.0].
-        validation_size
-            Size of the test set. If `None`, defaults to 1 - `train_size`. If
-            `train_size + validation_size < 1`, the remaining cells belong to a test set.
-        shuffle_set_split
-            Whether to shuffle indices before splitting. If `False`, the val, train, and test set
-            are split in the sequential order of the data according to `validation_size` and
-            `train_size` percentages.
-        batch_size
-            Minibatch size to use during training.
-        %(param_accelerator)s
-        %(param_devices)s
-        datasplitter_kwargs
-            Additional keyword arguments passed into
-            :class:`~scvi.dataloaders.SemiSupervisedDataSplitter`.
-        plan_kwargs
-            Keyword args for :class:`~scvi.train.SemiSupervisedTrainingPlan`. Keyword arguments
-            passed to `train()` will overwrite values present in `plan_kwargs`, when appropriate.
-        **trainer_kwargs
-            Other keyword args for :class:`~scvi.train.Trainer`.
-        """
-        if max_epochs is None:
-            max_epochs = get_max_epochs_heuristic(self.adata.n_obs)
-
-            if self.was_pretrained:
-                max_epochs = int(np.min([10, np.max([2, round(max_epochs / 3.0)])]))
-
-        logger.info(f"Training for {max_epochs} epochs.")
-
-        plan_kwargs = {} if plan_kwargs is None else plan_kwargs
-        datasplitter_kwargs = datasplitter_kwargs or {}
-
-        # if we have labeled cells, we want to subsample labels each epoch
-        sampler_callback = [SubSampleLabels()] if len(self._labeled_indices) != 0 else []
-
-        data_splitter = SemiSupervisedDataSplitter(
-            adata_manager=self.adata_manager,
-            train_size=train_size,
-            validation_size=validation_size,
-            shuffle_set_split=shuffle_set_split,
-            n_samples_per_label=n_samples_per_label,
-            batch_size=batch_size,
-            **datasplitter_kwargs,
-        )
-        training_plan = self._training_plan_cls(self.module, self.n_labels, **plan_kwargs)
-        if "callbacks" in trainer_kwargs.keys():
-            trainer_kwargs["callbacks"] + [sampler_callback]
-        else:
-            trainer_kwargs["callbacks"] = sampler_callback
-
-        runner = TrainRunner(
-            self,
-            training_plan=training_plan,
-            data_splitter=data_splitter,
-            max_epochs=max_epochs,
-            accelerator=accelerator,
-            devices=devices,
-            check_val_every_n_epoch=check_val_every_n_epoch,
-            **trainer_kwargs,
-        )
-        return runner()
 
     @classmethod
     @setup_anndata_dsp.dedent
