+++ src/scvi/model/base/__init__.py
@@ -10,7 +10,7 @@ from ._pyromixin import (
     PyroSviTrainMixin,
 )
 from ._rnamixin import RNASeqMixin
-from ._training_mixin import UnsupervisedTrainingMixin
+from ._training_mixin import SemisupervisedTrainingMixin, UnsupervisedTrainingMixin
 from ._vaemixin import VAEMixin
 
 __all__ = [
@@ -27,4 +27,5 @@ __all__ = [
     "JaxTrainingMixin",
     "BaseMinifiedModeModelClass",
     "EmbeddingMixin",
+    "SemisupervisedTrainingMixin",
 ]
